Update binaryConverter.java
Yes, the reason you're getting larger values than expected is because you're multiplying the ASCII values of the char instead of the int. There are multiple ways to work around this, one could be manually subtracting 48 from xx[i] (as the ASCII value of 0 is 48): dValue = (xx[I] - 48) * values[i]; or using Character.getNumericValue(xx[i]). Hope this helps!